OPL-1458 Wire up RoleAccessTable on Parameters Help Documentation

+++ src/components/inventory/parameters/ParametersInventory.jsx
@@ -11,7 +11,6 @@ import ParameterFilterModel from "components/inventory/parameters/parameter.filt
 import workflowAuthorizedActions
   from "components/workflow/pipelines/pipeline_details/workflow/workflow-authorized-actions";
 import InventorySubNavigationBar from "components/inventory/InventorySubNavigationBar";
-import ToolRegistryHelpDocumentation from "../../common/help/documentation/tool_registry/ToolRegistryHelpDocumentation";
 import ParametersHelpDocumentation from "../../common/help/documentation/tool_registry/ParametersHelpDocumentation";
 
 function ParametersInventory() {
@@ -79,7 +78,8 @@ function ParametersInventory() {
         parameters.forEach((parameter) => {
           const deleteAllowed = workflowAuthorizedActions.isActionAllowed(userRoleAccess, "delete_parameter", parameter.owner, parameter.roles, roleDefinitions);
           const updateAllowed = workflowAuthorizedActions.isActionAllowed(userRoleAccess, "update_parameter", parameter.owner, parameter.roles, roleDefinitions);
-          const newModel = {...new ParameterModel({...parameter}, newParameterMetadata, false, getAccessToken, cancelTokenSource, loadData, updateAllowed, deleteAllowed)};
+          const canEditAccessRoles = workflowAuthorizedActions.isActionAllowed(userRoleAccess, "edit_access_roles", parameter.owner, parameter.roles, roleDefinitions);
+          const newModel = {...new ParameterModel({...parameter}, newParameterMetadata, false, getAccessToken, cancelTokenSource, loadData, updateAllowed, deleteAllowed, canEditAccessRoles)};
 
           modelWrappedArray.push(newModel);
         });
@@ -94,6 +94,12 @@ function ParametersInventory() {
     }
   };
 
+  const getHelpComponent = () => {
+    if (!isLoading) {
+      return (<ParametersHelpDocumentation parameterRoleDefinitions={parameterRoleDefinitions} />);
+    }
+  };
+
   return (
     <ScreenContainer
       navigationTabContainer={<InventorySubNavigationBar currentTab={"parameters"} />}
@@ -101,9 +107,7 @@ function ParametersInventory() {
       pageDescription={`
         Parameters allow the user to store sensitive information in the vault in order to reference it later in the pipeline step.
       `}
-      helpComponent={
-        <ParametersHelpDocumentation />
-      }
+      helpComponent={getHelpComponent()}
     >
       <ParametersView
         isLoading={isLoading}
